Split remote daemon upload fallback helper
