+++ Packages/macOS/CmuxRemoteSession/Sources/CmuxRemoteSession/Session/RemoteSessionCoordinator.swift
@@ -569,32 +569,25 @@ public final class RemoteSessionCoordinator: @unchecked Sendable {
 
     // MARK: - Subprocess execution (through the runner seam)
 
-    func sshExec(
-        arguments: [String],
-        stdin: Data? = nil,
-        standardInputFile: URL? = nil,
+    func sshExec(arguments: [String], stdin: Data? = nil, standardInputFile: URL? = nil,
         timeout: TimeInterval = 15
     ) throws -> RemoteCommandResult {
         try runProcess(
             executable: "/usr/bin/ssh",
             arguments: arguments,
-            environment: configuration.sshProcessEnvironment,
-            stdin: stdin,
+            environment: configuration.sshProcessEnvironment, stdin: stdin,
             standardInputFile: standardInputFile,
             timeout: timeout
         )
     }
 
-    func scpExec(
-        arguments: [String],
-        timeout: TimeInterval = 30,
+    func scpExec(arguments: [String], timeout: TimeInterval = 30,
         operation: (any RemoteTransferCancelling)? = nil
     ) throws -> RemoteCommandResult {
         try runProcess(
             executable: "/usr/bin/scp",
             arguments: arguments,
-            environment: configuration.sshProcessEnvironment,
-            stdin: nil,
+            environment: configuration.sshProcessEnvironment, stdin: nil,
             timeout: timeout,
             operation: operation
         )
@@ -607,8 +600,7 @@ public final class RemoteSessionCoordinator: @unchecked Sendable {
         currentDirectory: URL? = nil,
         stdin: Data?,
         standardInputFile: URL? = nil,
-        timeout: TimeInterval,
-        operation: (any RemoteTransferCancelling)? = nil
+        timeout: TimeInterval, operation: (any RemoteTransferCancelling)? = nil
     ) throws -> RemoteCommandResult {
         try processRunner.run(
             RemoteProcessRequest(
